moved inject vote cache to post parser
