+++ src/providers/queries/postQueries/postQueries.ts
@@ -7,6 +7,7 @@ import { useAppSelector } from '../../../hooks';
 import { getDiscussionCollection, getPost } from '../../hive/dhive';
 import QUERIES from '../queryKeys';
 import { Comment, CacheStatus, LastUpdateMeta } from '../../../redux/reducers/cacheReducer';
+import { injectPostCache, injectVoteCache } from '../../../utils/postParser';
 
 /** hook used to return user drafts */
 export const useGetPostQuery = (_author?: string, _permlink?: string, initialPost?: any) => {
@@ -121,83 +122,14 @@ export const useDiscussionQuery = (_author?: string, _permlink?: string) => {
   );
 
   useEffect(() => {
-    _injectCache();
+    const _data =  injectPostCache(query.data, cachedComments, cachedVotes, lastCacheUpdate);
+    setData(_data);
   }, [query.data, cachedComments, cachedVotes]);
 
   useEffect(() => {
     restructureData();
   }, [data]);
 
-  // inject cached comments here
-  const _injectCache = async () => {
-    let shouldClone = false;
-    const _comments = query.data || {};
-    console.log('updating with cache', _comments, cachedComments);
-    if (!cachedComments || !_comments) {
-      console.log('Skipping cache injection');
-      return _comments;
-    }
-
-    // process votes cache
-    for (const path in cachedVotes) {
-      const cachedVote = cachedVotes[path];
-      if (_comments[path]) {
-        console.log('injection vote cache');
-        _comments[path] = _injectVoteFunc(_comments[path], cachedVote);
-      }
-    }
-
-    // process comments cache
-    for (const path in cachedComments) {
-      const currentTime = new Date().getTime();
-      const cachedComment = cachedComments[path];
-      const _parentPath = `${cachedComment.parent_author}/${cachedComment.parent_permlink}`;
-      const cacheUpdateTimestamp = new Date(cachedComment.updated || 0).getTime();
-
-      switch (cachedComment.status) {
-        case CacheStatus.DELETED:
-          if (_comments && _comments[path]) {
-            delete _comments[path];
-            shouldClone = true;
-          }
-          break;
-        case CacheStatus.UPDATED:
-        case CacheStatus.PENDING:
-          // check if commentKey already exist in comments map,
-          if (_comments[path]) {
-            shouldClone = true;
-            // check if we should update comments map with cached map based on updat timestamp
-            const remoteUpdateTimestamp = new Date(_comments[path].updated).getTime();
-
-            if (cacheUpdateTimestamp > remoteUpdateTimestamp) {
-              _comments[path].body = cachedComment.body;
-            }
-          }
-
-          // if comment key do not exist, possiblky comment is a new comment, in this case, check if parent of comment exist in map
-          else if (_comments[_parentPath]) {
-            shouldClone = true;
-            // in this case add comment key in childern and inject cachedComment in commentsMap
-            _comments[path] = cachedComment;
-            _comments[_parentPath].replies.push(path);
-            _comments[_parentPath].children = _comments[_parentPath].children + 1;
-
-            // if comment was created very recently enable auto reveal
-            if (
-              lastCacheUpdate.postPath === path &&
-              currentTime - lastCacheUpdate.updatedAt < 5000
-            ) {
-              console.log('setting show replies flag');
-              _comments[_parentPath].expandedReplies = true;
-              _comments[path].renderOnTop = true;
-            }
-          }
-          break;
-      }
-    }
-
-    setData(shouldClone ? { ..._comments } : _comments);
-  };
 
   // traverse discussion collection to curate sections
   const restructureData = async () => {
@@ -297,7 +229,7 @@ export const useInjectVotesCache = (_data: any | any[]) => {
 
       // if post available, inject cache and update state
       if (_postData) {
-        _postData = _injectVoteFunc(_postData, _voteCache);
+        _postData = injectVoteCache(_postData, _voteCache);
 
         if (_postIndex < 0) {
           console.log('updating data', _postData);
@@ -321,7 +253,7 @@ export const useInjectVotesCache = (_data: any | any[]) => {
         const _path = `${item.author}/${item.permlink}`;
         const voteCache = votesCollection[_path];
 
-        item = _injectVoteFunc(item, voteCache);
+        item = injectVoteCache(item, voteCache);
       }
       return item;
     };
@@ -332,28 +264,4 @@ export const useInjectVotesCache = (_data: any | any[]) => {
   }, [_data]);
 
   return retData || _data;
-};
-
-const _injectVoteFunc = (post, voteCache) => {
-  if (
-    voteCache &&
-    (voteCache.status !== CacheStatus.FAILED || voteCache.status !== CacheStatus.DELETED)
-  ) {
-    const _voteIndex = post.active_votes.findIndex((i) => i.voter === voteCache.voter);
-    if (_voteIndex < 0) {
-      post.total_payout += voteCache.amount * (voteCache.isDownvote ? -1 : 1);
-      post.active_votes = [
-        ...post.active_votes,
-        {
-          voter: voteCache.voter,
-          rshares: voteCache.isDownvote ? -1000 : 1000,
-        },
-      ];
-    } else {
-      post.active_votes[_voteIndex].rshares = voteCache.isDownvote ? -1000 : 1000;
-      post.active_votes = [...post.active_votes];
-    }
-  }
-
-  return post;
 };
